Some logic mistakes have been eliminated

+++ src/main/java/ru/javaops/topjava/web/vote/VoteController.java
@@ -3,32 +3,25 @@ package ru.javaops.topjava.web.vote;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.lang.Nullable;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-import ru.javaops.topjava.model.Meal;
 import ru.javaops.topjava.model.Restaurant;
 import ru.javaops.topjava.model.User;
 import ru.javaops.topjava.model.Vote;
 import ru.javaops.topjava.repository.RestaurantRepository;
 import ru.javaops.topjava.repository.VoteRepository;
-import ru.javaops.topjava.service.RestaurantService;
 import ru.javaops.topjava.service.VoteService;
-import ru.javaops.topjava.to.ResultTo;
 import ru.javaops.topjava.to.VoteTo;
-import ru.javaops.topjava.util.MealsUtil;
-import ru.javaops.topjava.util.UsersUtil;
 import ru.javaops.topjava.util.VoteUtil;
 import ru.javaops.topjava.web.AuthUser;
 
 import java.net.URI;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -37,7 +30,7 @@ import static ru.javaops.topjava.util.DateTimeUtil.atStartOfDayOrMin;
 import static ru.javaops.topjava.util.DateTimeUtil.atStartOfNextDayOrMax;
 import static ru.javaops.topjava.util.VoteUtil.createTo;
 import static ru.javaops.topjava.util.validation.ValidationUtil.assureIdConsistent;
-import static ru.javaops.topjava.util.validation.ValidationUtil.checkNew;
+
 
 @RestController
 @RequestMapping(value = VoteController.REST_URL, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -50,6 +43,8 @@ public class VoteController {
     private final RestaurantRepository restaurantRepository;
     private final VoteService service;
 
+    static final LocalTime DEADLINE = LocalTime.of(11, 0);
+
 
     @GetMapping("/{id}")
     public ResponseEntity<Vote> get(@AuthenticationPrincipal AuthUser authUser, @PathVariable int id) {
@@ -59,12 +54,19 @@ public class VoteController {
     }
 
     @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@AuthenticationPrincipal AuthUser authUser, @PathVariable int id) {
-        int userId = authUser.id();
-        log.info("delete {} for user {}", id, userId);
-        Vote vote = repository.getExistedOrBelonged(authUser.id(), id);
-        repository.delete(vote);
+    //@ResponseStatus(HttpStatus.NO_CONTENT)
+    public ResponseEntity<?> delete(@AuthenticationPrincipal AuthUser authUser, @PathVariable int id) {
+        if (LocalTime.now().isBefore(DEADLINE)) {
+            int userId = authUser.id();
+            log.info("delete {} for user {}", id, userId);
+            Vote vote = repository.getExistedOrBelonged(authUser.id(), id);
+            repository.delete(vote);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+        }
+
     }
 
     @GetMapping
@@ -91,18 +93,25 @@ public class VoteController {
     }
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Vote> createWithLocation(@AuthenticationPrincipal AuthUser authUser, @Valid @RequestBody VoteTo voteTo) {
+    public ResponseEntity<Vote> createWithLocation(@AuthenticationPrincipal AuthUser authUser, @Valid @RequestBody Integer restaurantId) {
         int userId = authUser.id();
-        log.info("create {} for user {}", voteTo, userId);
-        checkNew(voteTo);
-        Restaurant restaurant = restaurantRepository.getExisted(voteTo.getRestaurantId());
-        User user = authUser.getUser();
-        Vote vote = new Vote(restaurant, user, voteTo.getDateTime());
-        Vote created = service.save(userId, vote);
-        URI uriOfNewResource = ServletUriComponentsBuilder.fromCurrentContextPath()
-                .path(REST_URL + "/{id}")
-                .buildAndExpand(created.getId()).toUri();
-        return ResponseEntity.created(uriOfNewResource).body(created);
+        //int restaurantId = voteTo.getRestaurantId();
+
+
+        if (repository.getBetweenHalfOpenForUser(userId, atStartOfDayOrMin(LocalDate.now()), atStartOfNextDayOrMax(LocalDate.now())).isEmpty()) {
+            log.info("create {} for user {}", restaurantId, userId);
+            Restaurant restaurant = restaurantRepository.getExisted(restaurantId);
+            User user = authUser.getUser();
+            Vote vote = new Vote(restaurant, user, LocalDateTime.now());
+            Vote created = service.save(userId, vote);
+            URI uriOfNewResource = ServletUriComponentsBuilder.fromCurrentContextPath()
+                    .path(REST_URL + "/{id}")
+                    .buildAndExpand(created.getId()).toUri();
+            return ResponseEntity.created(uriOfNewResource).body(created);
+        }
+        else {
+            return new  ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+        }
     }
